Move relay responsibilities to submodules
